Created branch overlap and added paired-end read support to TSSV
TSSV will now do all alignments and pattern matching against an
explicitly uppercase copy of each read.  When flanking a flanking
sequence (as provided in the library file) has been detected in a
read, TSSV will check whether the section of the read that corresponds
to this flanking sequence.  If this section of the read was completely
lowercase, the match is rejected.

The idea behind this is that with paired-end data, it will be possible
to use a read merging tool to combine the two reads of each pair,
possibly giving rise to a longer combined read that contains a known
allele that did not fit completely in either read.  Because repetitive
sequences like STRs may be aligned in many ways, such tools will be
unable to detect whether the combined read actually contains all
repeats that were originally present in the DNA molecule.  If the
molecule was too long, the two reads may not overlap at all; however,
in a highly repetitive sequence, they may still appear to overlap
significantly.  It is only certain that all repeats have been captured
if all of them fitted on a single read - but the flanking sequences
may not be present in their entirety for such long alleles.  Merging
the two reads together may still yield the complete flanking sequence,
allowing the allele to be detected by TSSV.

The combined read, when provided to TSSV, should contain uppercase
letters for the overlapped part (i.e., the middle, where both reads in
the pair overlap) and lowercase letters for the non-overlapping parts
(i.e., those parts only present in one of the two reads).  If TSSV
aligns a flanking sequence to a completely lowercase part of the
combined read, the flanking sequence was only present in one of the two
reads and had fallen off the end of the other.  In this case, it is
impossible to tell whether the sequence between the flanking sequences
is complete and it is not meaningful to perform a regular expression
match against it.

Further modifications:
* Added -q option to read data from a FastQ file instead of FastA.